Workaround unreliable AUCell
We opt to explicitly coerce the count matrix to a dense matrix to avoid issues with support for sparse matrices in AUCell.

1. There are some changes to the outputs in the 'Cell type annotation' chapter that are independent of AUCell. These are minor and are fine to ignore.
2. There were changes made to AUCell in the BioC 3.15 release cycle that mean we now get different plots. Perhaps this is correcting previously wrong outputs, I don't now and don't have time to dig into. But this is independent of the issue with `AUCell::AUCell_buildRankings()`.
3. `AUCell::AUCell_buildRankings()` is unreliable.
  a. It gives different results for dense and sparse inputs.
  b. It does not automatically/gracefully handle sparse inputs as it should.

(2) is annoying, but for now we ignore it

(3a) is definitely a problem, but the AUCell authors don't seem inclined to fix this.
My assumption (hope?) is that the calculations for dense inputs are correct/reliable, so I'm inclined to use those.

(3b) is the killer and is what is preventing a clean build of OSCA.basic in both BioC 3.15 and 3.16.

See https://github.com/OSCA-source/OSCA.basic/issues/8 for further analysis and rationale